requirements.txt
Django==1.3.1
Pygments==1.4
distribute==0.6.19
django-guardian==1.0.3
psycopg2==2.4.4
wsgiref==0.1.2
